pass,pgcraft: change nb_vert_out_vars and nb_frag_output types to size_t

+++ libnopegl/src/node_render.c
@@ -190,7 +190,7 @@ static int render_init(struct ngl_node *node)
         .instance_attributes = o->instance_attributes,
         .nb_instances = o->nb_instances,
         .vert_out_vars = ngli_darray_data(&program_priv->vert_out_vars_array),
-        .nb_vert_out_vars = (int)ngli_darray_count(&program_priv->vert_out_vars_array),
+        .nb_vert_out_vars = ngli_darray_count(&program_priv->vert_out_vars_array),
         .nb_frag_output = program_opts->nb_frag_output,
         .blending = o->blending,
     };
